Add stress test and fix errors
Fixes duplicated message problem and refines resource closing
On close, the consumer would still iterate through all the messages it had in memory.
Reverts documentation to mention finer granularity topic
Fixes skipping messages on error cases.
If a consumer couldn't process a message for some reason, the whole batch
would be lost.
Since the message consumption should be ordered, we can seek the consumer
to the same offset on commit and ensure the next poll is on par with the server.